tests/wpt/web-platform-tests/html/semantics/embedded-content/the-embed-element/embed-document.html

<!doctype html>
<meta charset="utf-8">
<title>HTML Test: The embed element represents a document</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<meta name="assert" content="Check if the embed element represents a document when a text/html resource source is used">
<body>
  <script type="application/javascript">
   var childLoaded = false;
   t = async_test("Test document type embedding");
   addEventListener("load", t.step_func_done(function() { assert_true(childLoaded); }));
  </script>
  <embed src="embed-iframe.html">
</body>
